requirements.txt
# SpaceLink v4.1 Requirements
# Remote Desktop Solution with WebRTC

# ============ Core ============
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0
pydantic>=2.0

# ============ WebRTC ============
aiortc>=1.6.0
aiohttp>=3.8.0

# ============ Screen Capture ============
mss>=9.0.0
pillow>=10.0.0
numpy>=1.24.0

# ============ Input Control ============
pyautogui>=0.9.54
pynput>=1.7.6

# ============ Audio ============
sounddevice>=0.4.6
pyaudio>=0.2.13; platform_system == "Windows"

# ============ Clipboard ============
pyperclip>=1.8.2

# ============ System ============
psutil>=5.9.0
pywin32>=306; platform_system == "Windows"

# ============ Security ============
cryptography>=41.0.0
pyotp>=2.9.0

# ============ Hardware Encoding ============
# Note: FFmpeg must be installed system-wide

# ============ QR Code ============
qrcode>=7.4.0

# ============ Optional: AI Agent ============
# openai>=1.0.0  # Uncomment for AI features

# ============ Development ============
python-multipart>=0.0.6
aiofiles>=23.0.0
